Added editable grid to purchase details, removed it from the add purchase dialog because it became too cluttered.
