+++ Spbs.Main.WebUi/Pages/AddPurchaseDialog.razor
@@ -2,7 +2,7 @@
 @if (IsVisible)
 {
     <div class="modal fade show d-block" id="new-purchase-modal" tabindex="-1" role="dialog">
-        <div class="modal-dialog modal-lg" role="document">
+        <div class="modal-dialog" role="document">
             <div class="modal-content">
                 <div class="modal-header">
                     <h5 class="modal-title" id="titleLabel">New Purchase</h5>
@@ -12,8 +12,6 @@
                 </div>
 
                 <div class="modal-body">
-
-
                     <EditForm Model="@NewPurchase"
                               OnValidSubmit="@HandleValidSubmit"
                               OnInvalidSubmit="@HandleInvalidSubmit">
@@ -54,53 +52,6 @@
                         <button type="submit" class="btn btn-primary edit-btn">Submit</button>
                         <button type="button" class="btn btn-secondary" data-dismiss="modal" @onclick="@CloseDialog">Close</button>
                     </EditForm>
-
-                </div>
-                <div class="modal-footer">
-                    <RadzenButton Icon="add_circle_outline" style="margin-bottom: 10px" Text="Add New Product" Click="@InsertRow" Disabled=@(_purchaseItemToAdd != null) />
-                    <RadzenDataGrid @ref="_purchaseItemGrid" AllowFiltering="true" AllowSorting="true" EditMode="DataGridEditMode.Single"
-                                    Data="@NewPurchaseItems" TItem="NewPurchaseItemViewModel" RowUpdate="@OnUpdateRow" RowCreate="@OnCreateRow">
-                        <Columns>
-                            <RadzenDataGridColumn TItem="NewPurchaseItemViewModel" Property="ProductName" Title="Product Name" Width="150px">
-                                <EditTemplate Context="purchaseItem">
-                                    <RadzenTextBox @bind-Value="purchaseItem.ProductName" Style="width:100%; display: block" Name="ProductName"/>
-                                    <RadzenRequiredValidator Text="The product name is required" Component="ProductName" Popup="true"/>
-                                </EditTemplate>
-                            </RadzenDataGridColumn>
-                            <RadzenDataGridColumn TItem="NewPurchaseItemViewModel" Property="Price" Title="Price" Width="150px">
-                                <EditTemplate Context="purchaseItem">
-                                    <RadzenNumeric TValue="decimal" @bind-Value="purchaseItem.Price" RadzenTextBox Style="width:100%; display: block" Name="Price"/>
-                                    <RadzenRequiredValidator Text="Please provide a price" Component="Price" Popup="true" />
-                                    <RadzenNumericRangeValidator Min="0m" Text="Please provide a price that is not negative." Component="Price" Popup="true" />
-                                </EditTemplate>
-                            </RadzenDataGridColumn>
-                            <RadzenDataGridColumn TItem="NewPurchaseItemViewModel" Property="Description" Title="Description" Width="250px">
-                                <EditTemplate Context="purchaseItem">
-                                    <RadzenTextBox @bind-Value="purchaseItem.Description" Style="width:100%; display: block"/>
-                                </EditTemplate>
-                            </RadzenDataGridColumn>
-
-
-                            <RadzenDataGridColumn TItem="NewPurchaseItemViewModel" Context="purchaseItem" Filterable="false" Sortable="false" TextAlign="TextAlign.Center" Width="120px">
-                                <EditTemplate Context="purchaseItem">
-                                    <RadzenButton Icon="check" ButtonStyle="ButtonStyle.Primary" Class="m-1" Click="@((args) => SaveRow(purchaseItem))">
-                                    </RadzenButton>
-                                    <RadzenButton Icon="close" ButtonStyle="ButtonStyle.Light" Class="m-1" Click="@((args) => CancelEdit(purchaseItem))">
-                                    </RadzenButton>
-                                </EditTemplate>
-                            </RadzenDataGridColumn>
-                            <RadzenDataGridColumn TItem="NewPurchaseItemViewModel" Context="order" Filterable="false" Sortable="false" TextAlign="TextAlign.Center" Width="60px">
-                                <Template Context="purchaseItem">
-                                    <RadzenButton ButtonStyle="ButtonStyle.Danger" Icon="delete" Size="ButtonSize.Small" Class="m-1" Click="@(args => DeleteRow(purchaseItem))" @onclick:stopPropagation="true">
-                                    </RadzenButton>
-                                </Template>
-                                <EditTemplate Context="purchaseItem">
-                                    <RadzenButton ButtonStyle="ButtonStyle.Danger" Icon="delete" Size="ButtonSize.Small" Class="m-1" Click="@(args => DeleteRow(purchaseItem))">
-                                    </RadzenButton>
-                                </EditTemplate>
-                            </RadzenDataGridColumn>
-                        </Columns>
-                    </RadzenDataGrid>
                 </div>
             </div>
         </div>
