+++ Spbs.Main.WebUi/Pages/AddPurchaseDialog.razor.cs
@@ -19,7 +19,7 @@ public partial class AddPurchaseDialog
     
     public bool IsVisible { get; set; }
     public NewPurchaseViewModel NewPurchase { get; set; } = new();
-    public List<NewPurchaseItemViewModel> NewPurchaseItems { get; set; } = new();
+    
     
     protected override void OnInitialized()
     {
@@ -33,7 +33,7 @@ public partial class AddPurchaseDialog
         Guid userId = await _userService.GetLoggedInUserId();
         purchase.OwnerId = userId.ToString();
         
-        var response = await _mediator.Send(new InsertPurchase.Request(purchase));
+        var response = await _mediator.Send(new UpsertPurchase.Request(purchase));
         if (response.Success)
         {
             await NewPurchaseCallback.Invoke();
@@ -56,63 +56,5 @@ public partial class AddPurchaseDialog
         IsVisible = false;
         // Callback here 
         StateHasChanged();
-    }
-
-
-    private NewPurchaseItemViewModel? _purchaseItemToAdd { get; set; }
-    private RadzenDataGrid<NewPurchaseItemViewModel> _purchaseItemGrid { get; set; }
-    
-    async Task SaveRow(NewPurchaseItemViewModel purchaseItem)
-    {
-        if (purchaseItem == _purchaseItemToAdd)
-        {
-            _purchaseItemToAdd = null;
-        }
-
-        await _purchaseItemGrid.UpdateRow(purchaseItem);
-    }
-
-    void CancelEdit(NewPurchaseItemViewModel purchaseItem)
-    {
-        if (purchaseItem == _purchaseItemToAdd)
-        {
-            _purchaseItemToAdd = null;
-        }
-
-        _purchaseItemGrid.CancelEditRow(purchaseItem);
-    }
-
-    async Task DeleteRow(NewPurchaseItemViewModel purchaseItem)
-    {
-        if (purchaseItem == _purchaseItemToAdd)
-        {
-            _purchaseItemToAdd = null;
-        }
-
-        if (NewPurchaseItems.Contains(purchaseItem))
-        {
-            NewPurchaseItems.Remove(purchaseItem);
-            await _purchaseItemGrid.Reload();
-        }
-        else
-        {
-            _purchaseItemGrid.CancelEditRow(purchaseItem);
-        }
-    }
-
-    async Task InsertRow()
-    {
-        _purchaseItemToAdd = new NewPurchaseItemViewModel();
-        await _purchaseItemGrid.InsertRow(_purchaseItemToAdd);
-    }
-
-    void OnCreateRow(NewPurchaseItemViewModel newPurchaseItem)
-    {
-        Console.WriteLine("Inserted a new purchase item!");
-    }
-    
-    void OnUpdateRow(NewPurchaseItemViewModel newPurchaseItem)
-    {
-        Console.WriteLine("Updated a purchase item!");
     }
 }
